+++ src/shapes/Object/Object.ts
@@ -15,7 +15,7 @@ import type {
 } from '../../typedefs';
 import { classRegistry } from '../../util/class_registry';
 import { runningAnimations } from '../../util/animation/AnimationRegistry';
-import { clone } from '../../util/lang_object';
+import { cloneDeep } from '../../util/internals/cloneDeep';
 import { capitalize } from '../../util/lang_string';
 import { capValue } from '../../util/misc/capValue';
 import { createCanvasElement, toDataURL } from '../../util/misc/dom';
@@ -1365,7 +1365,7 @@ export class FabricObject<
           this._applyPatternForTransformedGradient(ctx, stroke);
         } else {
           // is a simple gradient or pattern
-          ctx.strokeStyle = stroke.toLive(ctx);
+          ctx.strokeStyle = stroke.toLive(ctx)!;
           this._applyPatternGradientTransform(ctx, stroke);
         }
       } else {
@@ -1378,7 +1378,7 @@ export class FabricObject<
   _setFillStyles(ctx: CanvasRenderingContext2D, { fill }: Pick<this, 'fill'>) {
     if (fill) {
       if (isFiller(fill)) {
-        ctx.fillStyle = fill.toLive(ctx);
+        ctx.fillStyle = fill.toLive(ctx)!;
         this._applyPatternGradientTransform(ctx, fill);
       } else {
         ctx.fillStyle = fill;
@@ -1584,7 +1584,7 @@ export class FabricObject<
       dims.zoomY / this.scaleY / retinaScaling
     );
     this._applyPatternGradientTransform(pCtx, filler);
-    pCtx.fillStyle = filler.toLive(ctx);
+    pCtx.fillStyle = filler.toLive(ctx)!;
     pCtx.fill();
     ctx.translate(
       -this.width / 2 - this.strokeWidth / 2,
@@ -1851,7 +1851,7 @@ export class FabricObject<
       ...options
     }: { extraParam?: string; signal?: AbortSignal } = {}
   ): Promise<FabricObject> {
-    return enlivenObjectEnlivables<any>(clone(object, true), options).then(
+    return enlivenObjectEnlivables<any>(cloneDeep(object), options).then(
       (enlivedMap) => {
         const allOptions = { ...options, ...enlivedMap };
         // from the resulting enlived options, extract options.extraParam to arg0