+++ src/shapes/line.class.ts
@@ -1,7 +1,6 @@
 import { SHARED_ATTRIBUTES } from '../parser/attributes';
 import { parseAttributes } from '../parser/parseAttributes';
 import { TClassProperties } from '../typedefs';
-import { clone } from '../util/lang_object';
 import { classRegistry } from '../util/class_registry';
 import { FabricObject, cacheProperties } from './Object/FabricObject';
 import { Point } from '../point.class';
@@ -102,7 +101,7 @@ export class Line extends FabricObject {
     // (by copying fillStyle to strokeStyle, since line is stroked, not filled)
     const origStrokeStyle = ctx.strokeStyle;
     if (isFiller(this.stroke)) {
-      ctx.strokeStyle = this.stroke.toLive(ctx);
+      ctx.strokeStyle = this.stroke.toLive(ctx)!;
     } else {
       ctx.strokeStyle = this.stroke ?? ctx.fillStyle;
     }
@@ -298,12 +297,16 @@ export class Line extends FabricObject {
    * @param {Object} object Object to create an instance from
    * @returns {Promise<Line>}
    */
-  static fromObject(object: Record<string, any>) {
-    const options = clone(object, true);
-    options.points = [object.x1, object.y1, object.x2, object.y2];
-    return this._fromObject(options, {
-      extraParam: 'points',
-    }).then((fabricLine) => {
+  static fromObject({ x1, y1, x2, y2, ...object }: Record<string, any>) {
+    return this._fromObject(
+      {
+        ...object,
+        points: [x1, y1, x2, y2],
+      },
+      {
+        extraParam: 'points',
+      }
+    ).then((fabricLine) => {
       return fabricLine;
     });
   }